feature/community Post 컴포넌트 나누기 및 게시물 좋아요 수정
